Python/tensorflow_experiments/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
decorator==5.1.1
dm-reverb==0.14.0
dm-tree==0.1.8
flatbuffers==23.5.26
gast==0.5.4
gin-config==0.5.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
gym==0.23.0
gym-notices==0.0.8
h5py==3.10.0
idna==3.6
imageio==2.4.0
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
portpicker==1.6.0
protobuf==4.23.4
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pygame==2.1.3
pyglet==2.0.10
PyVirtualDisplay==3.0
requests==2.31.0
requests-oauthlib==1.3.1
rlds==0.1.8
rsa==4.9
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflow-probability==0.23.0
termcolor==2.4.0
tf-agents==0.19.0
tf-keras==2.15.0
typing_extensions==4.5.0
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.14.1
